# fix delayed execution bug when using classic cache model