--- a/src/mem/cache/cache.cc
+++ b/src/mem/cache/cache.cc
@@ -807,7 +807,7 @@ Cache::serviceMSHRTargets(MSHR *mshr, const PacketPtr pkt, CacheBlk *blk)
                 // responseLatency is the latency of the return path
                 // from lower level caches/memory to an upper level cache or
                 // the core.
-                completion_time += curTick() + responseLatency +
+                completion_time += curTick() + responseLatency * clockPeriod() +
                     (transfer_offset ? pkt->payloadDelay : 0);
 
                 assert(!tgt_pkt->req->isUncacheable());