@giactra 2024-11-27
Thanks @2channelkrt for this fix. Can I ask you to split the changes in two separate PRs?

The CHI change definitely needs to go as soon as possible in next release