@giactra 2024-11-27
> > Thanks @2channelkrt for this fix. Can I ask you to split the changes in two separate PRs?
> > The CHI change definitely needs to go as soon as possible in next release
> 
> It's two commits. Is that ok?
> 
> I can cherry-pick the CHI one when merged

I haven't read the problem description in full, but the classic cache change looks wrong to me.

`completion_time += curTick() + responseLatency`

Here we are adding Tick to cycles. This is wrong. clockEdge was rightfully transforming the latency value (cycles) into the corresponding tick number. So that's why I believe we should split this in two; the CHI change is straightforward and fixes CHI in the staging branch.

We need extra time to review the problem description for the classic bit