@2channelkrt 2024-11-28
Hi @BobbyRBruce and @giactra! thanks for the review.

> Here we are adding Tick to cycles. This is wrong. clockEdge was rightfully transforming the latency value (cycles) into the corresponding tick number.

Original code was transforming latency into clockEdge and queueing data-ready instructions to mark them ready, and this can be seen as a right way to go, but 'marking the instruction ready' and 'executing CPU tick' was done in different order between 'classic cache' and 'ruby cache' caused the delay. So queueing data-ready instrucitons at tick rather than clockEdge solves this issue for most case.

For the case where cache line is ready on the cpu clockEdge, issue would still occur, as they will be ready on clockEdge even if we use tick() to queue instructions to mark them ready.