@giactra 2024-11-28
> Hi @BobbyRBruce and @giactra! thanks for the review.
> 
> > Here we are adding Tick to cycles. This is wrong. clockEdge was rightfully transforming the latency value (cycles) into the corresponding tick number.
> 
> Original code was transforming latency into clockEdge and queueing data-ready instructions to mark them ready, and this can be seen as a right way to go, but 'marking the instruction ready' and 'executing CPU tick' was done in different order between 'classic cache' and 'ruby cache' caused the delay. So queueing data-ready instrucitons at tick rather than clockEdge solves this issue for most case.
> 
> For the case where cache line is ready on the cpu clockEdge, issue would still occur, as they will be ready on clockEdge even if we use tick() to queue instructions to mark them ready.

The fact that ruby and classic caches' behaviour differ does not surprise me. Ultimately it's a matter of understanding which model resembles the operation of a real world cache. To this I have little knowledge but it does not look unreasonable to me to respond to multiple memory operation targeting the same line in the same cycle

However it looks to me you are just tampering the event scheduling to make things work. But effectively it is just a coincidence and you are doing so adding two values which should not be added together.

Look at the snipper I posted before:

`completion_time += curTick() + responseLatency`

responseLatency is defined as being number of Cycles. You cannot add Cycles to Ticks (the output of curTick()); that 's like adding apples and pears.

Tick is the atomic unit of measure of gem5 time. Everything is scheduled according to ticks. Cycles are made of a number of ticks (>1) and that number depends on the frequency of the object (you can have multiple clock domains)

So for example

[ _ _ _ _ _ _ _ _ _ _] = 10 Ticks

[ _ _ ] = 1 Cycle = 2 Ticks for a 2 GHz CPU

[ _ _ _ _ ] = 1 Cycle = 4 Ticks for a 1 GHz CPU

Ok, so the moment a user defines responseLatency as number of Cycles, you cannot directly add that value to a Tick number because it will not give the expected results. In the above case, say you model an operation that will take 2 cycles.

[ _ _ _ ] = 1 Cycle = 3 Ticks

x = curTickTime();

Simulation time: [ _ _ x _ _ _ _ _ _ _ _ _ _ _ _ _]

If you want the operation to happen in 2 cycles, you cannot do

y = x + 2
Simulation time: [ _ _ x _ y _ _ _ _ _ _ _ _ _ _ _]

You need to do

y = x + clockEdge(2) = x + 3 * 2 = x + 6
Simulation time: [ _ _ x _ _ _ _ _ y _ _ _ _ _ _ _]

Hope this helps.