@2channelkrt 2024-11-28
I got the point. When i tested it with resp latency 1 it worked as intended (1 cpu clock cycle, 400ticks for 2.5Ghz) but i will double check with other values. 

Even if other value works as intended, adding tick and latency without any explicit conversion seem absurd to me too, this i will add some conversion method for resp latency for better readability and make sure the value is synched to the clock.

will that suffice?