@2channelkrt 2024-11-29
i confirmed the code is wrong.
when curtick() is 500 and resplatency is 5 while cpu clock is 2.5GHz, my intended comp_time is 500 + (5 * 400) = 2500 but now it's printing 505. I will push fix for this.