@2channelkrt 2024-11-29
fixed. this is what I intended to do.