@giactra 2024-11-29
> fixed. this is what I intended to do.

The new version is effectively similar to [clockEdge](https://github.com/gem5/gem5/blob/stable/src/sim/clocked_object.hh#L183)

    Tick
    clockEdge(Cycles cycles=Cycles(0)) const
    {
        // align tick to the next clock edge
        update();

        // figure out when this future cycle is
        return tick + clockPeriod() * cycles;
    }

The only difference is the update method which is aligning the computation starting point to the next clock edge, if we are not (curTick) on a clock edge. This makes sense as it assumes changes (triggering an elaboration) will be made visible to the logic only on a clock edge (and that should be the starting count on the latency).

I fail to understand the logical connection between this line change and what you are effectively trying to do. Unless somehow calling ClockedObject::update over different MSHR targets yields to different completion times.
But as far as I can see from the code, this shouldn't be the case.

I believe you should make sure (debugging is a solution) the PR is not fixing your problem coincidentally/indirectly. The starting point shouldn't be: I want the CPU to get the data at this tick time and I am going to scram the timing of the cache model so that this condition is satisfied. The starting point should rather be: I want to be able to service multiple targets at the same cycle. Why is this not happening?

Also, no need to act on this now, but it would be nice if you could squash all your commits into a single one and force push the PR ;)